\name{demoMemUsage1}
\alias{demoMemUsage1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Code that show memory used and available
}
\description{
Code that show memory used and available
}
\usage{
demoMemUsage1()
}
\value{
To be ignored
}
\author{
Nick Sabbe
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
demoMemUsage1()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ memory }
\keyword{ nspack }% __ONLY ONE__ keyword per line
