\name{nspack-package}
\alias{nspack-package}
\alias{nspack}
\docType{package}
\title{
Utility functions only useful on computers accessible to Nick Sabbe
}
\description{
Utility functions only useful on computers accessible to Nick Sabbe
}
\details{
\tabular{ll}{
Package: \tab nspack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-24\cr
License: \tab GNU\cr
LazyLoad: \tab yes\cr
}
	Internal \enumerate{
		\item \code{\link{demoMemUsage1}}
		\item \code{\link{doSf}}
	}
}
\author{
Nick Sabbe

Maintainer: <nick.sabbe@ugent.be>
}
\keyword{ package }
\examples{
demoMemUsage1()
}
